Atualização da Documentação (Correção de BPMN's) - Funções de Projetos e usuários

+++ back-end/src/controllers/usuarios.js
@@ -1,9 +1,24 @@
-// Validar a senha do usuário, ao tentar alterar os dados ou exlcuir a conta
+// Importando bibliotecas necessárias
 import prisma from '../database/client.js';
 import bcrypt from 'bcrypt';
 
 const controller = {};
 
+// Função para validar senha do usuário
+async function validaSenha(senhaAtual, idUsuario){
+    const verificaUsuario = await prisma.usuario.findUnique({
+        where: { id: idUsuario }
+    }); 
+    
+    let verSenha = await bcrypt.compare(senhaAtual, verificaUsuario.senha);
+    if(verSenha){
+        return true;
+    }else{
+        return false;
+    }
+}
+
+// Validado (12-04) - Implementar Cadastro de Imagens
 // Criando um novo usuário
 controller.create = async function(req, res) {
     try {
@@ -23,8 +38,22 @@ controller.create = async function(req, res) {
 
         await prisma.usuario.create({ data: req.body });
 
-        // Retornando mensagem de sucesso
-        res.status(201).json({mensagem: "Usuário Cadastrado com Sucessso!"});
+        // Recuperando os dados do recem criado usuário para iniciar a sessão no vanegador
+        const usuarioCadastrado = await prisma.usuario.findUnique({
+            where: { email: req.body.email }
+        });
+
+        // Criando a sessão no navegador
+        req.session.usuario = {
+            id_usuario: usuarioCadastrado.id,
+            nome_usuario: usuarioCadastrado.nome,
+            email_usuario: usuarioCadastrado.email,
+            foto_usuario: usuarioCadastrado.foto
+        };
+
+        // Retornando resultado para redirecionamento no front
+        // console.log(req.session.usuario);
+        res.status(201).json({result: true});
     }
     catch(error) {
         // Deu errado: exibe o erro no terminal
@@ -36,6 +65,8 @@ controller.create = async function(req, res) {
     }
 }
 
+// Desativa após o desenvolvimento
+// Não se faz necessário no momento, e poderia ser perigoso tal ferramenta ser usada de forma incorreta
 // Obtendo todos os usuários cadastrados
 controller.retrieveAll = async function(req, res) {
     try {
@@ -63,6 +94,7 @@ controller.retrieveAll = async function(req, res) {
     }
 }
 
+// Validado (12-04)
 // Obtendo um usuário específico pelo id
 controller.retrieveOne = async function(req, res) {
     try {
@@ -90,16 +122,41 @@ controller.retrieveOne = async function(req, res) {
     }
 }
 
-// Obtendo um usuário específico pelo id
-controller.retrieveOneEmail = async function(req, res) {
+// Validado (12-04)
+// Obtendo um usuário específico pelo email (Login)
+controller.loginEmail = async function(req, res) {
     try {
         // Buscando os dados de um usuário específico
-        const result = await prisma.usuario.findUnique({
+        const verificaEmail = await prisma.usuario.findUnique({
             where: {email: req.params.email}
         });
     
-        // Retorna os dados obtidos
-        res.send(result);
+        if (!verificaEmail){
+            return res.status(400).json({mensagem: "Usuário Não Encontrado!"});
+        }
+
+        // Verificando se a senha informada é a senha atual do usuario
+        const valSenha = await validaSenha(req.body.senha_atual, verificaEmail.id);
+        if (!valSenha){
+            return res.status(400).json({ mensagem: "Senha Inválida!"});
+        }
+
+        // Recuperando os dados alterados do usuário para iniciar a sessão no vanegador
+        const usuarioCadastrado = await prisma.usuario.findUnique({
+            where: { email: req.params.email }
+        });
+
+        // Alterando os dados da sessão
+        req.session.usuario = {
+            id_usuario: usuarioCadastrado.id,
+            nome_usuario: usuarioCadastrado.nome,
+            email_usuario: usuarioCadastrado.email,
+            foto_usuario: usuarioCadastrado.foto
+        };
+
+        // Retornando resultado para redirecionamento no front
+        // console.log(req.session.usuario);
+        res.status(201).json({result: true});
     }
     catch(error) {
         // P2025: erro do Prisma referente a objeto não encontrado
@@ -117,6 +174,7 @@ controller.retrieveOneEmail = async function(req, res) {
     }
 }
 
+// Validado (12-04) - Implementar Cadastro de Imagens novas e remover as antigas
 // Atualizando os dados do usuário
 controller.update = async function(req, res) {
     try {
@@ -137,6 +195,12 @@ controller.update = async function(req, res) {
             }
         }
 
+        // Verificando se a senha informada é a senha atual do usuario
+        const valSenha = await validaSenha(req.body.senha_atual, req.params.id);
+        if (!valSenha){
+            return res.status(400).json({ mensagem: "Senha Inválida!"});
+        }
+        delete req.body.senha_atual;
 
         // Criptografando a senha antes de ser alterados os dados
         const senhaCriptografada = await bcrypt.hash(req.body.senha, 10);
@@ -148,8 +212,21 @@ controller.update = async function(req, res) {
             where: { id: req.params.id },
             data: req.body
         });
+
+        // Recuperando os dados alterados do usuário para iniciar a sessão no vanegador
+        const usuarioCadastrado = await prisma.usuario.findUnique({
+            where: { email: req.body.email }
+        });
+
+        // Alterando os dados da sessão
+        req.session.usuario = {
+            id_usuario: usuarioCadastrado.id,
+            nome_usuario: usuarioCadastrado.nome,
+            email_usuario: usuarioCadastrado.email,
+            foto_usuario: usuarioCadastrado.foto
+        };
     
-        // Retrnando mensagem de sucessao caso tenha atualizado
+        // Retornando mensagem de sucessao caso tenha atualizado
         res.status(201).json({mensagem: "Usuário Atualizado com Sucesso!"});
     }
     catch(error) {
@@ -169,16 +246,33 @@ controller.update = async function(req, res) {
     }
 }
 
+// Validado (12-04) - Implementar remoção de imagens do usuário
 // Deletando o usuário
 controller.delete = async function(req, res) {
     try {
-      // Busca o usuário a ser excluído
-      await prisma.usuario.delete({
-        where: { id: req.params.id }
-      });
-  
-      // Envia mensagem confirmando a exclusão
-      res.status(201).json({mensagem: "Usuário Deletado com Sucessso!"});
+        // Verificando se a senha informada é a senha atual do usuario
+        const valSenha = await validaSenha(req.body.senha_atual, req.params.id);
+        if (!valSenha){
+            return res.status(400).json({ mensagem: "Senha Inválida!"});
+        }
+        delete req.body.senha_atual;
+
+        // Busca o usuário a ser excluído
+        await prisma.usuario.delete({
+            where: { id: req.params.id }
+        });
+
+        // Destruindo sessão
+        req.session.destroy(err => {
+            if (err) {
+            res.status(500).json({ mensagem: 'Erro ao encerrar sessão' });
+            } else {
+            console.log('Logout realizado com sucesso');
+            }
+        });
+    
+        // Envia resultado confirmando a exclusão
+        res.status(201).json({result: true});
     }
     catch(error) {
       // P2025: erro do Prisma referente a objeto não encontrado
@@ -197,4 +291,4 @@ controller.delete = async function(req, res) {
     }
   }
 
-export default controller
+export default controller;
